+++ src/modules/esm.mjs
@@ -1,0 +1,39 @@
+import path from "path";
+import { release, version } from "os";
+import { createServer } from "http";
+import "./files/c.js";
+
+const random = Math.random();
+
+let unknownObject;
+if (random > 0.5) {
+  unknownObject = import("./files/a.json", {
+    assert: { type: "json" },
+  });
+} else {
+  unknownObject = import("./files/b.json", {
+    assert: { type: "json" },
+  });
+}
+
+const filename = import.meta.url;
+const dirname = path.dirname(filename);
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${filename}`);
+console.log(`Path to current directory is ${dirname}`);
+
+const myServer = createServer((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+export { unknownObject, myServer };
